+++ kernel/core/init.c
@@ -6,8 +6,6 @@
 #include <linux/sched.h>
 #include <linux/workqueue.h>
 #include <linux/moduleparam.h>
-#include <linux/rwsem.h>
-#include <linux/utsname.h>
 
 #include "policy/allowlist.h"
 #include "policy/app_profile.h"
@@ -27,6 +25,7 @@
 #include "hook/syscall_hook.h"
 #include "feature/adb_root.h"
 #include "feature/selinux_hide.h"
+#include "feature/uts_spoof.h"
 #include "infra/symbol_resolver.h"
 
 #if defined(__x86_64__)
@@ -125,51 +124,17 @@ int __init kernelsu_init(void)
         pr_alert("shell is allowed at init!");
     }
 
-    if (spoof_release || spoof_version) {
-        struct rw_semaphore *sem = (struct rw_semaphore *)find_kernel_symbol_exact("uts_sem");
-        struct uts_namespace *ns = (struct uts_namespace *)find_kernel_symbol_exact("init_uts_ns");
-
-        if (sem) {
-            down_write(sem);
-        } else {
-            down_write(&uts_sem);
-        }
-
-        if (ns) {
-            if (spoof_release) {
-                strscpy(ns->name.release, spoof_release, sizeof(ns->name.release));
-            }
-            if (spoof_version) {
-                strscpy(ns->name.version, spoof_version, sizeof(ns->name.version));
-            }
-        } else {
-            if (spoof_release) {
-                strscpy(init_uts_ns.name.release, spoof_release, sizeof(init_uts_ns.name.release));
-            }
-            if (spoof_version) {
-                strscpy(init_uts_ns.name.version, spoof_version, sizeof(init_uts_ns.name.version));
-            }
-        }
-
-        if (sem) {
-            up_write(sem);
-        } else {
-            up_write(&uts_sem);
-        }
-
-        if (ns) {
-            pr_info("ksu: spoofed version: %s, release: %s\n", ns->name.version, ns->name.release);
-        } else {
-            pr_info("ksu: spoofed version: %s, release: %s\n", init_uts_ns.name.version, init_uts_ns.name.release);
-        }
-    }
-
     ksu_cred = prepare_creds();
     if (!ksu_cred) {
         pr_err("prepare cred failed!\n");
     }
 
     ksu_init_symbol_resolver();
+
+    if (spoof_release || spoof_version) {
+        ksu_spoof_version(spoof_release, spoof_version);
+    }
+
     ksu_syscall_hook_init();
 
     ksu_feature_init();
